requirements.txt
fastapi==0.73.0
uvicorn[standard]==0.17.4
celery==5.2.3
flower==1.0.0
httpx ==0.21.1
python-multipart





-f https://download.pytorch.org/whl/cpu/torch_stable.html
numpy
scipy
scikit-learn
matplotlib
pandas
gunicorn
python-multipart
uvloop
httptools
torch==1.10.2+cpu
torchvision==0.11.3+cpu
torchaudio==0.10.2+cpu
opencv-python-headless
# opencv-python
# mmcv-full
mmcv-full==1.4 -f https://download.openmmlab.com/mmcv/dist/cpu/torch1.10.0/index.html


######mmdet

matplotlib
numpy
pycocotools
six
terminaltables


######mmpose

chumpy
dataclasses; python_version == '3.7'
json_tricks
matplotlib
munkres
numpy
pillow
scipy
xtcocotools>=1.8


######mmseg

matplotlib
mmcls>=0.20.1
numpy
packaging
prettytable


######DB

tortoise-orm==0.19.2
asyncpg